reduce BFG bright flash fix time
Just skip drawing the projectile only when the BFG is fired (one frame). Adapted to the new md3 model. It should work well with any g_delagMissileBaseNudge setting. With workaround turned on, it looks a bit better, although there is not that much of a difference. The number of milliseconds during which BFG projectile frames will not be drawn can be increased by changing the value of cg_ratFixPredictedBFG. Set to 0 to disable.